requirements.txt
flask
<<<<<<< HEAD
psycopg2-binary
pandas
google-auth
openpyxl
=======
gunicorn
psycopg2-binary
google-auth
requests
pandas
>>>>>>> 482e1205a295fd1055e3568df4bf7514753257e9
